Added a WorkflowIOHang exception class that will be raised whenever a IO hang is detected.
Changed the WorkflowIOProxy to raise the WorkflowIOHang exception when hangs are detected
and when access to paths known to cause hangs is attempted.

Fixed some bugs in the book keeping of paths that cause hangs.

Added code so that WorkflowIOHang exceptions are caught and handled appropriately when
they occur during checking for resolution of file dependencies.

Need to add code to catch WorkflowIOHang exceptions anywhere calls to the WorkflowIOProxy 
are made.