Added separate stop! method definition in the dbproxy to prevent it from restarting the DB server just so that it can then shut it down in the event that the server has died prior to the shutdown attempt.
Reworded some text in exception handling for stop! methods such that the user is informed that
the server had died prior to the stop, but that it is not going to restart it just so that it
can shut it down.  Instead the error is noted because something went wrong, but ignored because
there is nothing to do to recover from it.