This is the first step toward hardening the batch system server interaction.
The ProxyBatchSystem class was renamed BQS because it is no longer the proxy.

A BQSProxy class was created to field batch system requests and forward them
to the BQS class.  The BQSProxy handles exception handling for the BQS server
and will automatically restart the server if it dies.

The BQS class is responsible for doing multithreaded job 
submission and for querying the batch system.  It uses the appropriate
class corresponding to the batch system used to make those queries.

Other parts of the code were modified to accommodate the new class names.

This commit does NOT add code to deal with missing job ids that result when
the batch system server dies.