This is the second step of having the bqserver record jobids as soon as they are returned by the qsub command.
A new database table was created to manage the URIs of active
bqserver processes.  Entries are added when a new bqserver
is launched.  Entries are removed when a bqserver is successfully
shutdown.  A bqserver is shut down if all threads running qsub
have terminated and all output from the qsub commands has either
been retrieved by the main WFM process or recorded in the DB.

A mutex is used in the bqserver to synchronize threaded access
to the DB so that only one thread will try to record a jobid 
at any given time.

Multiple retries and random sleeps are used to deal with DB 
access collisions between the main WFM thread and the bqserver.
While this is not ideal, it works in practice because the 
number of collisions is usually very small.  There is room 
for improvement in the handling (or prevention) of these collisions.