This is the first cut at implementing support for the moabtorque scheduler option. This is initially targeted for use on Zeus, but it should work anywhere that uses MOAB on top of Torque.  Jobs are submitted and deleted using the Torque qsub and qdel commands.  Jobs are monitored and tracked via the MOAB showq command.
This relies on job completion information persisting for a least long enough for
the WFM to retrieve job status information after a job terminates.  If the 
MOAB keep time is too short, the WFM won't be able to get job completion 
information and will fail.

The logging has also been fixed such that jobs which are deleted will be 
reported to have a run time of 0 instead of a gigantic integer value.