Add support for a <deadline> tag for <task> tags.  This allows more flexibility in expiring tasks in situations where some tasks for a cycle might need to expire at different times. The cyclelifespan attribute still works well for cases when all tasks for a cycle expire simultaneously.
Fix bugs in comparisions of cycles.

Add automatic capping of wallclock time requests for tasks such that the submission time
plus the wall clock time requested is never greater than the time when the cycle or task expires.

Fix minor bug in display of wfmstat to report exit status as "-" unless the job actually finished.

Update version to 1.0.1