Bug fix: If get_compound_time_string is passed a nil argument, a nil is now returned without attempting to treat it as an element.
Bug fix?: Added calls to start the garbage collector as instructed in the libxml-ruby documentation to
prevent xpath objects from being freed before the document object during abnormal termination of the code.
According to the documentation, seg faults can result in that scenario due to the way libxml2 memory is
managed and the fact that Ruby doesn't guarantee any kind of order of freeing objects in that situation.