Bug fix and feature addition:
A bug causing rocotocheck to crash with undefined method for nil was fixed.  This bug would be triggered if rocotocheck
was run for a task that was defined but did not have any dependencies (a rare thing).

A feature was added to rocotocheck to print the status of hang dependencies as well as normal dependencies.