Bug Fix:
The previous attempt to avoid zombie processes by detaching from child processes does
not work.  The problem is that methodology only works if the parent process always lives
longer than the child process.  And this is not the case here.  

Here we restructure how Rocoto "server" processes are launched to properly avoid the
accumulation of zombie processes.  The "server" processes now immediately  daemonize
themselves such that they are completely decoupled from the process that launched them.
This greatly complicated the tracking of pids of the server processes and necessitated
changes to the way abandoned processes are automatically shutdown when they are abandoned.
It also required changes to the pids used to control workflow locking.

In addition, a logging method is implemented to record events that are important to 
track but that do not belong in the workflow logs.

Note:  rocotobqserver processes are still not automatically shutdown if they are
abandoned.  This is tricky and has not been solved yet.