Bug Fix:
On some systems it appears that there is a race condition or a 
/tmp filesystem coherency problem that causes the URI file that 
is written by the server to have partial, or no content when
the client that reads it detects that it exists.  This results
in occassional failures to read the URI file, which results in 
inability to connect to the server process that was launched,
which results in the server process being abandoned.

An additional polling loop was added to make sure the entire 
line containing the URI written by the server is read before
attempting to connect to the server process using the URI
that was read.