Bug Fixes:
When the Moab/Torque batch queueing system goes down or becomes unresponsive, the 
inability to retrieve job status information was resulting in a task state of
"Unknown".  This is undesirable, because even though the state is unknown, it is
unknown because the batch system could not be queried and not because the status
could not be found.  This results in many tasks exceeding the maximum unknown 
count during batch outages, which causes erroneous submissions of jobs that have
completed normally.  There is no point in attempting to query a down batch system.

The fixes in this commit change the behavior such that a failure to query the batch
system results in a task state of "Unavailable" rather than "Unknown".  There is no
limit to the number of times a task can be in the "Unavailable" state.  A state of
"Unknown" now only results when all queries to the batch system succeed but the job
status can not be found.  A state of "Unknown" should now be confined to cases where
a bug in Moab/Torque results in the job being lost, or there is a delay in Moab 
obtaining job information from Torque (a delay of a couple minutes is normal).  

Additionally, commands used to query the batch system for job status will now be
killed if they do not finish within a deadline.  Any failure of any command
to query the batch system due to hangs or errors will cause the batch system
to be marked as being down.  If a batch system is marked as down, no further 
attempts to query the batch system will be made until the next time Rocoto is
run, and all job status will be reported as "Unavailable" for the current iteration.

Tasks and workflow progress should no longer be penalized for a batch system outage.

A utility for writing messages to STDERR according to verbosity settings is
also added.