In order to log Rocoto events to STDERR according to the verbosity level set by the -v option in the various rocoto programs, the verbosity setting must be passed to many different classes and/or methods.  This has become onerous, especially since other options, such as the database file and workflow document file also need to be passed.
In this commit the Rocoto main programs set the Rocoto options and Rocoto
configuration as constants in the WorkflowMgr module.

WorkflowMgr::OPTIONS
WorkflowMgr::CONFIG

This enables all parts of the code to reference these values very easily.  While 
I am not a huge fan of this type of design, as it defeats the purpose of OO design
principles, I can't think of a more elegant solution at the moment.  These 
values need to be accessed by almost every class, so this is the easiest
way to accomplish this right now.  Passing references to the option classes to
every class that needs them was too messy and time consuming at this juncture.

This will be revisited in the future once a proper solution has been found.

NOTE: This commit does NOT enable the setting of WorkflowMgr::OPTIONS and 
WorkflowMgr::CONFIG inside the code accessed by rocoto server processes.
That requires a different approach.