Bug fixes and enhancements:
This is a large commit that contains a number of important updates.

1) The method for logging Rocoto events has been updated.  

Important events that are not specific to a particular cycle, but which
relate to system errors are always logged in $HOME/.rocoto/log.  This
facilitates diagnosis of problems and errors even when users are not
enabling logging themselves or are redirecting cron output to /dev/null.

Events which go to $HOME/.rocoto/log and other events that have a
corresponding cycle associated with them will also now go to STDERR if
the user has set the verbosity of the rocoto commands accordingly.  Very
important events will be written to STDERR regardless of verbosity.  Most
other notifications will now only be written if verbosity > 0.

A lot of code was touched in the process of implementing this mechanism.

This will continue to evolve.  The verbosity is read from ARGV and set
in WorkflowMgr::VERBOSE.  WorkflowMgr.stderr and WorkflowMgr.log 
write the messages as appropriate for the verbosity specified.



2) A lot of simplification has been made to the main programs in /sbin.
   Code was reorganized for efficiency and ease of maintenance.

3) Major improvements to the moabtorque and torque batch system classes
   was made for increased reliablity in the event of batch system failures.
   The occurrence of the notorious "UNKNOWN" state of tasks is greatly 
   reduced.  In the event of a Moab or Torque outage, the state will now
   be reported as "UNAVAILABLE".  There is no limit to the number of times
   a job can be in the "UNAVAILABLE" state, but once the batch system 
   returns to service, an updated state will be reported.  The "UNKNOWN"
   state will still occur if the batch system is fully functional AND the
   job state is not found.  In the case of moabtorque, if MOAB is up and
   is unable to find the status of the job, Torque is queried directly.
   If neither MOAB or Torque find the job, and they are both up, the state
   will be set to UNKNOWN.