Bug Fix:
When an XML file is parsed it may contain external entities
which are references to other files that must be parsed
in addition to the top level file.  The method used to parse
XML files did not take that into consideration.  While the
main file was read by the workflowIOServer process, the
parsing of the data was done inside the main Rocoto process
using the string parser.  Thus, any files associated with
external entities would be read by the main Rocoto process.

Another problem, found and reported by Tim Brown, is that
the string parser does not always work properly when the
string contains references to external entities.

This has been rectified.  Now the XML files for the XML
workflow definition as well as the RelaxNG validation
file are parsed by the workflowIOServer process using the
file parser isntead of the string parser.  The resulting
document is not marshalable across a DRb connection.  So 
the parsed document, which has had full entity replacement
done, is converted to its string representation and sent
back to the Rocoto main process.  The main process then
uses the string parser to parse the document.  Since no
entities are present in the string, the bug reported by
Tim is avoided.

This fix does result in slower code as the document is now
parsed twice.  Unfortunately, no way around that has been
found which ensures that all file IO takes place in the
workflowIOServer process.  Performance still seems reasonable
so far.