Bug Fix:
The Rocoto bqserver process stores results of job submission commands
that it issues until a rocotorun or rocotoboot process retrieves them.
In certain situations it is possible for the bqserver process to be
forgotten and abandoned.  The automated shutdown of the bqserver 
process isn't possible until we find way for the bqserver process to
deliver its data directly to the database as soon as it is available.
There are a lot of issues with doing that.  For now, we trigger an
automatic shutdown of the bqserver process if it has been alive for
more than 4 days.  This allows manual runs of rocotorun that spawn
bqserver processes to work across weekends without disrupting the
workflow.  This time may be shortened later.