Feature modification:
Most of the standard messages written to stderr that inform
about various types of failures are only written at verbosity
1 and above.  The default verbosity is modified to 1 so that
the default behavior is to get those messages.  This will 
maintain default behavior of previous releases.  Those messages
can be turned off by setting the verbosity of the rocoto programs
to 0.