Bug Fix:
The Libxml-ruby library automatically sends warnings, errors, and fatal errors to
STDERR, whether you want it to or not.  Even if you handle the exceptions it raises
for those errors it still sends the output there.

This has very negative consequences for tools that are launched from cron where all
output needs to be carefully controlled to avoide unwanted emails from crond.

This commit installs a so-called "handler" to turn off this ridiculous behavior.

Messages about errors in your XML document should still get sent to STDERR by Rocoto.
However, messages about errors in Torque and Moab XML output (due to errors produced by
those tools in certain scenarios) will NOT go to STDERR unless it is appropriate for 
them to go there.

Enhancement:

A new utility, run4, for running commands with separate file descriptors for stdin,stdout,
and stderr is added.  The run4 utility will replace the run utility in commits to follow.