Bug Fix:
When Torque and/or Moab are configured in an HA mode and one of the HA pairs is down,
calls to showq and/or qstat may display warning messages to stdout in addition to
the normal expected output that goes to stdout. For example,

[harrop@e0285 test]$ qstat
Error communicating with ebqs1(10.176.113.24)
Job id                    Name             User            Time Use S Queue
------------------------- ---------------- --------------- -------- - -----
12007.ebqs1                test2            harrop          00:00:00 C batch          
12008.ebqs1                test1            harrop          00:00:00 C batch          
12010.ebqs1                test4            harrop          00:00:00 C batch          
12009.ebqs1                test3            harrop          00:00:00 C batch          
12012.ebqs1                test5            harrop          00:00:00 C batch          
12013.ebqs1                test2            harrop          00:00:00 C batch          

Since Rocoto uses the XML output of these tools, and the output that goes to 
stderr is NOT written in XML format, this caused a problem because the 
commands were run with stderr redirected to stdout.  The XML parser would
generate a fatal exception due to malformed XML.

The error message notwithstanding, in this situation, the batch system is considered 
UP because the HA node that is UP is still handling batch system requests without 
problems.  So it is important to parse the normal output and ignore the error in
this scenario.

This commit uses the new run4 utility to seperate stdout and stderr and to only send the
stdout to the XML parser.  The parser will now only throw an error if the output of
batch system qstat and showq commands that goes to STDOUT is malformed (a genuine error).
If a down situation is detected the output of the commadn that went to STDERR will be 
logged and optionally written to Rocoto's STDERR.