Bug Fix:
When crond forks a process, that process, and all processes that it forks inherits stdout and stderr
from crond.  If the process that crond forks is outlived by child processes that it forks,
then the process that crond forks will stick around as a zombie if any of it's children are using 
the stdout or stderr file descriptors.  This commit adds a safety thread that checks to see
if the parent has become a zombie process.  As soon as the parent is a zombie, the child closes 
stderr, allowing the parent to die and removing the zombie.