Behavior modification:
Moved the checking of dependencies when considering tasks for submission such
that the check happens prior to the check for throttle violations.  Previously,
the throttle violations were checked first because those checks are computationally
cheap compared to the checking of dependencies.  This allowed the checking of 
dependencies to be skipped in cases where throttles would be violated.  However,
this caused confusing log messages because throttle violations were reported 
(when applicable) for tasks that had no chance to run because their dependencies
were not yet satisfied.  This commit removed the confusion such that throttle
violations will only be reported for tasks whose dependencies are satisfied.
There is a performance penalty in some workflows for doing this, but it should
be pretty small.