Bug Fix (maybe):
After doing thousands of runs of WRF NMM Rocoto test workflows and having a user run the
Rocoto trunk for several days, a total of 4 segmentation faults was observed while running
rocotorun.  All of them occurred in the same place.  All of them occurred on Zeus.

/home/Christopher.W.Harrop/rocoto-trunk/lib/workflowmgr/utilities.rb:217: [BUG] Segmentation fault
ruby 1.8.7 (2011-06-30 patchlevel 352) [x86_64-linux]

This commit attempts to address this very random and very rare (so far) crash that has
thus far been impossible to reproduce.  The potential fix is based on information found at:

http://tickets.opscode.com/browse/CHEF-2916

It seems that it is possible that Ruby attempts to allocate objects during GC when calling
IO.select and IO#read which leads to the segmentation faults.  This commit simply disables
garbage	  collection during calls to the run4 utility which calls   the open4 utility which	
manages	  the IO pipes.  Only time will tell whether or not this fixes the problem.