Bug Fix:
In extremely rare situations when the Rocoto server processes are forked
the main thread can die before the child call Process.ppid.  This results
in the child's ppid being assigned to init (process 1). This, in turn,
can result in a rogue (not a zombie) process that never goes away.  By
making sure that the ppid is set before the fork, even if the ppid is
dead by the time the server starts, it will be able to shut itself down
cleanly.