Restrict recording of backtraces to exceptions that occur due to ArgumentError, NameError, and TypeError.
This isn't perfect, but we really don't want a backtrace for 
every exception that occurs because many of them are handled
properly with an appropriate message that is printed and/or logged.

We really only want backtraces for Rocoto bugs, not for when
Rocoto properly handles a system or user error by raising an
exception.  This can probably be improved.