Behavior modification:
This commit introduces a new task state called "LOST".  A task will be assigned to the 
LOST state if the maximum number of consecutive UNKNOWN states has been reached.  For 
all intents and purposes, the LOST state is treated the same as the FAILED state.  The 
purpose of distinguishing the two is to better convey to the user what happened to the
task.  FAILED implies that the job ran but crashed, which is not necessarily the case
for a job whose state was irretrievably lost due to a batch system problem.  LOST means
that the state could not be found and that the job may or may not have run correctly.
LOST tasks are resubmitted just as FAILED tasks are in order to err on the side of 
maintaining workflow integrity.