Bug Fix:
The handling of locking with regard to opening the database in dbopen
was not being properly handled.  The dbopen was erroneously being retried
in the event that it raised a SQLite3::BusyException.  While that does 
serialize access to the database to prevent corruption, it should not 
happen.  A side effect of this was the recording of many messages about
the database being locked, when only one message was necessary.  This
commit fixes this behavior by disabling retries for the dbopen method.
If a dbopen fails, the rocoto process will write a single message saying
that the database is locked and then quit.