Bug Fix from Sam Trahan:
The parsing of bhist output did not allow for use of certain characters like "." 
in usernames, hostnames, and queue names.  This caused the parsing of bhist output
to fail and certain job features could not be captured.  This bug was not seen
until Sam installed Rocoto on WCOSS Tide.  Many thanks to Sam for finding the 
bug and sending the patch.