Bug Fix for problem reported by Michelle Harrold
If a task entered the DEAD state but was removed from the
workflow definition, either before or after becoming DEAD,
Rocoto was erroneously continuing to write log messages to
report the DEAD state forever.  Once a task goes DEAD, no 
further logging for it should take place unless the task is 
resurrected by the user increasing the maxtries attribute
for the task.  This simple fix ensures that only one
log entry is made to report a DEAD task that is no longer
defined.