Bug Fix:
Many thanks to Tim Brown for this brilliant idea!

This commit eliminates the use of /tmp (or any filesystem) for retrieving
the URI of DRb rocoto server processes.  In the process, it fixes bugs
related to the accidental reuse of /tmp/rocoto_uri_* files that caused
file permission problems and sometimes resulted in Rocoto trying to 
connect to the wrong URI.  It also fixes a bug that resulted in the 
workflowIOProxy looking at the wrong rocotoioserver PID when attempting
to recover from IO hangs (the daemonization process resulted in the use
of the wrong pid).

The main Rocoto process now opens a pipe before forking the processes
that daemonize themselves and become the rocoto*server processes.  The
children inheret the file descriptors for these pipes, so they can be
used on both ends to communicate the URI and the eventual server PID.
The forked server process simply opens the inherited write end of the pipe
and writes the URI and its PID (it simply closes the read end of the pipe).
Meanwhile the main Rocoto process reads the read end of the pipe, and 
retrieves the URI and the pid, enabling it to connect to the DRb Rocoto
server process (it simply closes the write end of the pipe).

Thus, the filesystem is no longer used for communications between processes.

This does need to be more rigorously tested under a more realistic workflow
load and adverse operating conditions, but it is working reliably in a
light testing environment.