Bug Fix:
This update should fix the reporting of job status for the scenario
where Torque is working and showq commands are returning valid
results without errors but MOAB is stuck an is not receiving updated
job informaiton from Torque.  Previously it was believed that the
Torque qstat command would be intrusive and many of them might harm
Torque server performance.  This is, apparently, not the case.  The
order of queries has thus been rearranged.  Now the Torque qstat 
command is tried first.  If the job information is not found, then
"showq -c" is queried.  As before, if the commands fail the state
will be reported as "UNAVAILABLE".  A state of "UNKOWN" can only 
occur if both Torque and Moab appear to be functioning and the job
information cannot be found.

This commit should prevent jobs from entering the "UNKNOWN" state
if they happen to finish while Torque is working, but MOAB is 
hung as described above as long as the cron interval for 
rocotorun is less than the job state keep time in the Torque config.