Bug Fixes and Enhancements:
This fixes a problem where multiple messages about database locking
failures were printed to stderr and/or $HOME/.rocoto/log.  Now, only
a single message per locking failure is written.

In addition to the above, the exception handling and error reporting
for all Rocoto server processes was revamped to improved the clarity
of messages and to eliminate unnecessary duplicates.

The locking of the database was modified to allow any call to retry
in the event of a SQLite locking error.  The algorithm for calculating
intervals between retries was modified to be more timely and to 
reduce collisions.  This allows retrieval of information about the 
owner of the database lock that can then be reported.  Otherwise, 
that information would not be available if an attempt to lock or open
the database was flat out rejected on a SQLite DB locking failure.