Bug Fix:
Implemented hash and eql? methods for CompoundTimeString and CycleString objects so that 
two equivalent CompoundTimeString objects will be considered equal when being inserted
into a hash.  The latest insertion of a value for a given key will overwrite any previous
value for that key.  This means, that if a user has:

      <envar>
        <name>START_TIME</name><value><cyclestr>@Y@m@d@H</cyclestr></value>
      </envar>
      <envar>
        <name>START_TIME</name><value><cyclestr>dummy</cyclestr></value>
      </envar>
      <envar>
        <name>START_TIME</name><value><cyclestr>dummy2</cyclestr></value>
      </envar>

The START_TIME will be set to dummy2.