Branch commit:
Add ability to assign names to metatasks in preparation for metatask dependencies.
Add first implementation of metatask throttles.

The "throttle" attribute for <metatask> allows users to control the maximum number of
tasks enclosed within it that may be queued or running at any given time.  This is 
generally more useful than simple individual task level throttling because it can 
throttle instances of a group of tasks intead of instances of individual tasks.

For example, if some post tasks were created using a <metatask>, the metatask 
throttling can limit the total number of post tasks that can be active at one time,
whereas the individual task throttle and the global workflow task throttle can not.