Branch commit:
Add initial implementation of metatask dependencies.
It has been lightly tested.

This implementation converts all <metataskdep> tags
into the corresponding 

<and>
  <taskdep ..../>
  <taskdep .../>
  ...
</and> 

sequences.  The cycle_offset and state attributes
are supported.  However, it does not yet handle the 
option for specifying a threshold of metatask members 
that must succeed or fail.  That will be in the next 
commit, and will use <some> instead of <and>.