Branch commit:
Add "threshold" attribute to <metataskdep> to enable specification
of a percentage of metatask items that must be satisfied.

The insertion of an <and> clause was replaced by a <some> clause
and the "threshold" attribute of the <some> is simply set to
1.0 as a default (when the <and> behavior is desired) if it is
not specified.

Still only very lightly tested.  Which is why this is a branch!