Branch Commit:
Add initial implementation of task geometry for Torque and LSF.

The <nodes> tag is added, which contains a Moab/Torque type nodespec
describing the desired task geometry.  However, the ppn spec is 
REQUIRED, and there is an additional and optional tpp spec that can
be used when MPI/OpenMP hybrid modes are being used.

For example,

<nodes>1:ppn=1+10:ppn=12</nodes>

The above requests one core on the first node, and 12 cores on the
last 10 nodes.

Another example:

<nodes>10:ppn=4:tpp=3</nodes>

The above requests 4 cores on each of 10 nodes, where each of those
cores will spawn 3 threads, consuming a total of 12 cores per node.

For Moab/Torque, the syntax in the <nodes> tags is passed in verbatim 
after removing occurrences of the tpp resource request (not needed, 
and not supported)

For LSF, it is considerably more complicated.  But, the <nodes> 
is translated into the appropriate "-R span[ptile=X]" and "-n Y"
options.  Plus the LSF PBL_TASK_GEOMETRY environment variable is
set to reflect the task geometry request.

This is currently VERY lightly tested, but should not be too far
from the mark.