BRANCH Commit:
The XML Error handler is changed so that it does not suppress the informative information
about XML validation errors.  There is still some work to be done as the line numbers
in the output are wrong.  Doing "xmllint file.xml | head -LINENO" will get the correct line
as a workaround.