BRANCH Commit:
This is the first cut at implementing automatic serialization of <metatask> contents.

A new metatask attribute is introduced that allows users to declare that the contents of
the given metatask are to be run sequentially.  The attribute is used as follows:

  <metatask name="foo" mode="serial">

The "mode" attribute is optional.  Only two values are allowed, "serial" and "parallel".
If the "mode" attribute is not set, the default is "parallel".

For <metatasks> that are declared to be "serial", the appropriate <taskdep> task dependencies
are automatically inserted on <metatask> expansion such that the immediate children of that
<metatask> will be run sequentially.  If one of the immediate children of the serial <metatask>'s 
is itself a <metatask>, the contents of that child metatask will not be serialized, but all of
it's children must run before the next sibling of the serial metatask can run.  Serialization
of all descendents can be accomplished by declearing all descending metatasks to be "serial".

THIS IS ONLY LIGHTLY TESTED AT THIS POINT!