BRANCH Commit:
The various messages that Rocoto outputs to stderr have been stratified into
categories representing a progression of fatal errors that one would always 
want to see to informational/debugging messages.

Also, one bug was fixed where an argument was missing for a method call in
the delete method of the the moabtorquebatchsystem class.