BRANCH COMMIT:
Bug Fix:  Many thanks to Javier Delgado for finding and reporting this bug.

An extraneous "-v" option was being inserted into the qsub command for tasks
which did not contain any <envar> tags.  This caused the qsub command not to 
parse properly.  This bug was introduced when a feature was added to replace 
-v with -V if a long -v string caused the command to be too long.

It should now be fixed.