BRANCH COMMIT
Small update to the test XLM and script

Modified the output to stdout such that notification of DEAD tasks happens at -v=1
and notification of task crashes happens at -v=3.

Updated libxml2 to version 2.9.1.

Updated libxml-ruby to version 2.7.0

Updated INSTALL script to install above versions of libxml2 and libxml-ruby.

Added calls to @workflowdoc.import as part of an attempt to avoid libxml-ruby seg faults

Added calls to set the results of XPath find commands to nil followed by a call to 
start the garbage collector.  This, in theory, helps prevent libxml-ruby seg faults.