BRANCH COMMIT:
I just tagged a release candidate, so of course I had to find another
problem.

It appears that in certain cases, garbage collection is needed in the
method that does metatask depenency expansion.  This is similar to
the situation in the method that does metatask serialization.

I believe the problem is caused by data structures in the LibXML
being reused and pointers going to the wrong place.