Bug Fix:
This commit fixes the notorious bug that causes this message to be
printed in .rocoto/log logs.

 Fatal error: Extra content at the end of the document at :2

The root cause was a limit built in to the libxml2 library that limits
the size of XML input that can be parsed.  If the input XML is larger than
that limit, it truncates the input and attempts (but fails) to parse 
what it got.  This would not happen for a Rocoto XML workflow definition, 
as those files are well below the limit.  But, the output of Moab's 
"showq -c --xml" command and Torque's "qstat -x" command are often (but 
unpredictably) very long.  The parsing of the output of these commands
would fail, and tracking of jobs would be hobbled.

The fix is to add an option to the parser telling it to not impose that
limit and allow the whole input string to be parsed.

Due to the potential size of the output of those commands, an alternative
approach for obtaining that information should probably be considered.