Torque behavior workaround:
Because of the way Torque's qsub command parses command line arguments, the
-v option does not handle the passing of environment variables whose value 
contains spaces, commas, or minus signs.  Setting the values of the variables
in the current environment and then using -V does work.

This commit allows Rocoto to prefer the use of -v when submitting jobs.  However,
if any variable's value (as specified in <envar> tags) contains a space, comma,
or minus sign, it will now automatically switch to the -V method and will not
use -v.