New Feature:
This is the initial implementation of a new feature that allows uses to declare
an arbitrary set of tasks to be "final" tasks.  Once a "final" task for a given
cycle completes successfully, the state of that cycle is set to "draining".  A
Rocoto will continue to track queued or running jobs for a "draining" cycle,
but no new jobs for that cycle will be submitted, and no failed jobs for that 
cycle will be resubmitted.  Once there are no queued or running jobs for a 
"draining" cycle, the cycle's state is changed to "done" regardless of the state
of any other task.  This feature allows cycles to age off even if some tasks 
have dependencies that will never be satisfied.  This is useful in cases where
a set of tasks will never be executed due to dependencies that will never be
satisfied because of a condition that is not known until runtime.  The aging
of these cycles once a "final" task is completed, allows cycle throttling to
work properly.

To use this feature, users simply add the "final" attribute to the task(s) the
wish to be "final" and set its value to true (T, t, True, TRUE, true).

<task name="mytask" ..... final="T" ... >

There is no limit on the number of tasks that can be "final" tasks.

Minimal testing has been done successfully.  However, more testing is needed 
to ensure that this feature works as intended.