Added a new <cycleexistdep> tag that can be used to have a task depend on the existence of a specific cycle in the XML file.  This is intended to be used for special actions to be taken at the first cycle of a run, or after intentional breaks in cycling, without having to use file or time manipulation tricks.
Example:

  <dependency>
    <or>
      <taskdep task="launch" cycle_offset="-6:00:00"/> 
      <not>
        <cycleexistdep cycle_offset="-6:00:00"/> 
      </not>
    </or>
  </dependency>

That will run if the prior cycle "launch" job was complete, or if the
prior cycle does not exist.  Note that one must be careful with this
feature since a new cycle may pop up in the middle of a storm in
real-time.

In the source code, the main difference is that the caller now has to
pass an array of CycleDef to all Dependency.resolved? and
Dependency.query calls.