Many updates to rocoto, some big, some small.  "To do" list at the end.  In no particular order:
1. JOB SUBMISSION CHANGE: On Torque and Moab+Torque, the submission is
now done by a script sent to the qsub stdin.  This is to work around a
critical problem on Jet: the <nodes> tag (and -l nodes=) does not work
on the command line.  This change was tested on Jet and Zeus with a
variety of job cards, and seems to work.

2. BUG FIX: When a metatask does not have a metatask throttle, the
workflow manager would exit with a confusing message about comparison
to nil.  The workflow manager now 

3. BUG FIX: When &quot; is used in a <command>, it is retained as
&quot; because it is impossible to turn off output escaping in LibXML
Ruby (!?). I added an "unescape" function that replaces a few common
XML entities with their equivalents (&lt; &gt; &amp; &quot;).  

4. NEW DEPENDENCIES:

   <true>, <false>
     Always true or false.  Use these to enable or disable parts of
     the workflow.

   <streq><left>string1</left><right>string2</right></streq>
   <strneq><left>string1</left><right>string2</right></strneq>
     String comparison "dependencies."  The first is string equality,
     and the second is for string inequality.  These are also for
     turning on or off parts of the workflow.  (For example,
     &RUN_OCEAN; compared to YES for the coupled forecast job, and NO
     for an uncoupled forecast job.)

   <rb name="depname">ruby script</rb>
     Evaluates a Ruby expression in a boolean context.  The result is
     the dependency.  There are a variety of local variables defined
     in a (somewhat) sandboxed environment, for accessing information
     about the task and cycle.  See WorkflowMgr::StringEvaluator for a
     full list.
     TO DO: Need to add something to catch common exceptions and treat
      them as the dependency failing.

   <sh>shell command</sh>
     Evaluates a shell command as a dependency.  If the result is 0,
     the dependency is met, otherwise it is not met.
     TO DO: Need to add something to catch failures to execute the
       program, and treat it as the dependency not being met.

5. NEW TASKDEP STATE: You can now trigger a job when another job is in
"Running" state.  This is intended to be used for jobs that can (but
don't have to) run in parallel with one another.  Note that ONLY
triggering based on "Running" is unwise since the job may complete
before Rocoto checks it.  You should always use this in an <or> block,
to test if the job is done or running.

6. REWIND FUNCTIONALITY: There is a new rocotorewind script, which
erases all information about specific tasks for specific cycles, hence
causing Rocoto to try running them again.  In addition, it can run
<rewind> scripts, which are <rb> and <sh> blocks.  
  <task ...>
     ... stuff ...
     <rewind>
       <sh>rm -f &WORK;/flagfile*</sh>
     </rewind>
  </task>
In the above example, when that task is rewound, all flagfile* files
in the &WORK; directory are deleted.
    
7. CODE SIMPLIFICATION: In the WorkflowMgr, there were blocks of code
repeated many times.  Adding the rewind functionality would have
repeated it yet again.  They have been wrapped up into a
"with_locked_db" function, that locks the database, runs the specified
code, and then unlocks the database.  This does not change the
workflow manager functionality in any way: it just simplifies the code
slightly.

8. WORKFLOWSTATE - There is a new WorkflowState class, that is passed
to all Dependency-like objects, when running query, resolved? or the
new rewind! functions.  This contains the old arguments sent to those
functions, plus the new ones required by the <rb>, <sh> and
<cycleexists> tags.  


########################################################################
============================= TO DO LIST ===============================
########################################################################

1. Add error checking to the <sh> tag so it will treat a failure to
run the program as "false" instead of terminating rocoto.

2. Add error checking to the <rb> tag so it treats exceptions as
"false" instead of terminating rocoto, EXCEPT for syntax errors.

3. Rewind cannot handle cycles that are expired (past a time limit) or
drained (ones that completed a "final" state task).  Not handling
expired cycles is fine, but a drained cycle must be rewindable, if the
final state tasks are all rewound.  

4. Remove any stray puts and other debug junk.